adding deprecated message every time print is run
