+++ src/cmdstan/print.cpp
@@ -3,6 +3,7 @@
 #include <iomanip>
 #include <ios>
 #include <stan/mcmc/chains.hpp>
+#include <cmdstan/print_helper.hpp>
 #include <cmdstan/stansummary_helper.hpp>
 #include <fstream>
 
@@ -19,6 +20,7 @@ int main(int argc, const char* argv[]) {
   
   if (argc == 1) {
     print_usage();
+    print_deprecated();
     return 0;
   }
   
@@ -36,6 +38,7 @@ int main(int argc, const char* argv[]) {
     
     if (std::string("--help") == std::string(argv[i])) {
       print_usage();
+      print_deprecated();
       return 0;
     }
     
@@ -52,6 +55,7 @@ int main(int argc, const char* argv[]) {
   
   if (!filenames.size()) {
     std::cout << "No valid input files, exiting." << std::endl;
+    print_deprecated();    
     return 0;
   }
   
@@ -314,10 +318,6 @@ int main(int argc, const char* argv[]) {
     }
         
   }
-      
+  print_deprecated();  
   return 0;
-        
 }
-
-
-
